pyproject.toml
[project]
name = "depth-estimation"
version = "0.1.0"
description = "Monocular metric depth estimation with depth prior and INR calibration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "datasets>=2.16.0",
    "huggingface_hub>=0.20.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "h5py>=3.8.0",
    "tqdm>=4.66.0",
    "scikit-image>=0.21.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["depth_estimation"]


[tool.ruff]
line-length = 100
target-version = "py310"
